# common: use unordered_map instead of map in cache