Search complexity.

For map, it is O(lgN)
For unordered_map, it is O(1) 

Signed-off-by: Xinze Chi <xmdxcxz@gmail.com>