@tchaikov 2015-04-23
@XinzeChi xinze, could you put the rationale in the commit message? why unordered_map is better than map in this case? for example, its impacts on performance? thanks.