@tchaikov 2015-04-23
thanks @XinzeChi for explaining the reason in the comment. i meant you could update it in the commit log. but i get your point. the `std::unordered_map` has its advantage over `std::map` in some cases. but the cache could be fairly large and sparse. in that case, the rb tree would perform better, i guess. and `std::unordered_map` could suffer from the rehash and bad distribution of inserted elements.

anyway, it's just my two cents.