@dachary 2015-05-13
@XinzeChi looks good to me. Since lower_bound is only used on the weak_refs, order does not matter. Could you please rebase and repush ? Just so that the bot compiles against the current master. After it confirms,
<pre>
Reviewed-by: Loic Dachary &lt;ldachary@redhat.com>
</pre>