@tchaikov 2015-05-13
@dachary yes, the order of std::map is a bonus feature we don't care, but its performance is very stable. unordered_map has its downsides as a hashmap as listed in my last comment. that's why i was not 100% sure about this pull request. anyway, i am good if you think it's a safe change.