@dachary 2015-05-13
@tchaikov I don't have actual experience with that but I'd be surprised if it's doing worse, overall. And I can't think of a situation where the hash rehash time would create a problem. 

The worst possible case would be that a thread is blocked long enough to trigger timeouts or a race condition that was not previously discovered. Is it what you also had in mind ?