@gregsfortytwo 2015-05-13
Woah, no way guys. Hash maps take up *way* more space than maps do, and we've had memory issues in the past from using hash_maps on a per-PG basis (like we use the SharedLRU). And as a rule we don't merge performance-only patches (especially ones with performance tradeoffs!) without performance data.

If we have performance tests indicating that the use of a regular map is an issue for some code, then we can look at changing things (probably on a configurable basis so we only use it for the hot one). But the usual size of these caches is on the order of 1024 entries, and that doesn't take many instructions to descend to the bottom.