@dachary 2015-05-13
@XinzeChi maybe you have performance analysis results that suggest this pull request would significantly improve things ?