@mslovy 2015-12-11
Hi all. @XinzeChi  @tchaikov  @gregsfortytwo @dachary. Any evidence indicates that unordered_map really cost much more memery, and impact performance because of rehash?
Based on my knowledge,  bucket acts as a slot, but actually allocate memory when real data inserted? I not sure, but under my test, when half of elements erased, buckets_size remains the same but memory is already free. 
Also, no evidence shows unordered_map costs more memory than map, but save insertion and searching times?

For map:
memery init--> %CPU:0.0	                MEM:2572KB
cost for 100000 insert: 57334800ns
memory after 100000 insert-->               MEM:16764KB
memory after 50000 items erased-->      MEM:9900KB
cost for 200000 insert: 114140824ns
memory after 200000 insert-->           MEM:30884KB

For unordered_map:
memery init-->                                    MEM:2612KB
cost for 100000 insert: 30892844ns
memory after 100000 insert-->             MEM:16396KB
memory after 50000 items erased-->   MEM:10092KB
cost for 200000 insert: 49562857ns
memory after 200000 insert-->             MEM:29760KB 

Am I right, or missing something?
test program in attachment and change it to .cc,  with complie flag -std=c++0x -ltcmalloc -lrt
[map_test.txt](https://github.com/ceph/ceph/files/59346/map_test.txt)

On the other side, FDcache is no need to use get_next and low_bound, so weak_ref also cannot be used as unordered_map. And with my test, save the lookup() execute time as below:

Searching speed:
current map<>
cache_size     shard      average cost
5120             128          1.74us
5120             10            2.43us
    
with unordered_map<>
cache_size     shard      average cost
5120             128          1.21us
5120             10            1.39us

1)save cpu about 0.2% under all-hits cases, will save more under low hit rates and,
2)release the ratio (cache_size:shard) to make cache more effective