@liewegas 2015-12-11
IIRC the memory explosion we observed was ages ago with hash_map from old versions of the STL; perhaps it is better now.

Also, the issue was that even a small/sparse map used lots of memory.  We have a SharedLRU for every PG with just 64 items.  The SharedLRU knows the expected size of the map... if that can be passed into the unordered_map to inform the number of buckets then I think we're fine (this is true with boost::intrusive::unordered_set, but I haven't looked at unordered_map).