@mslovy 2015-12-11
Yes, we can call rehash(int bucket_number) to pre-set the bucket number. 
refer to http://www.cplusplus.com/reference/unordered_map/unordered_map/rehash/
I think this kind of work would make sense now? wait others comment