@liewegas 2015-12-11
That's perfect. If the SharedLRU etc constructor calls that to size the bucket array accordingly that should work great!